# PERF: CategoricalDtype.__eq__